+++ src/Libraries Projects/Milochau.Core.AspNetCore/Infrastructure/Features/ConfigurationBuilderService.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Configuration;
+using Milochau.Core.Abstractions;
 using Milochau.Core.AspNetCore.Models;
-using Milochau.Core.Models;
 using Milochau.Core.AspNetCore.Infrastructure.Middlewares;
 using Microsoft.AspNetCore.Routing;