+++ src/Libraries Projects/Milochau.Core.AspNetCore/Infrastructure/Features/TelemetryBuilderService.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
+using Milochau.Core.Abstractions;
 using Milochau.Core.AspNetCore.Models;
-using Milochau.Core.Models;
 
 namespace Milochau.Core.AspNetCore.Infrastructure.Features
 {