+++ src/Reference Projects/Milochau.Core.AspNetCore.ReferenceProject/Pages/Index.cshtml.cs
@@ -1,5 +1,5 @@
-﻿using Milochau.Core.AspNetCore.ReferenceProject.Models;
-using Milochau.Core.Models;
+﻿using Milochau.Core.Abstractions;
+using Milochau.Core.AspNetCore.ReferenceProject.Models;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Options;
 using Microsoft.FeatureManagement;